requirements.txt
djangorestframework >= 3.14
requests >= 2.31.0
retrying